Double in Data

+++ Source/Components/Loyalty/Broadcaster/MLBusinessLoyaltyBroadcastData.swift
@@ -12,7 +12,7 @@ public class MLBusinessLoyaltyBroadcastData: NSObject {
     public let percentage: Float
     public let primaryColor: String
 
-    public init(level: Int, percentage: Double, primaryColor: String) {
+    public init(level: Int, percentage: Float, primaryColor: String) {
         self.level = level
         self.percentage = percentage
         self.primaryColor = primaryColor
